requirements.txt
# requirements.txt
kafka-python
sseclient
requests
firebase
python_jwt
gcloud
pycrypto
requests_toolbelt
SPARQLWrapper
